Replaced left-right padding with container
Instead of controlling the left and right layout of the main sections with the right adn left padding, we now control it with a container that has a set width. We did this chaneg because it's going to be easier to modify the right and left layout for larger screens, and for better code maintenance.