Modified animations effects
Modified hero animation effect. Now, instead of zooming-into each individual element, I packed the animation of the title and contact button together, to avoid code pollution. Also, that fixed the hover/focus state of the contact button not working. Plus, I added an offset of 350px in the initialization of each animations, to make the animations appear whelement are at 250px of the bottom of the screen.